[carver] delete tools with empty check (#384)

+++ carver/plugins/tool-repository-postgresql/src/main/java/modelengine/jade/carver/tool/repository/pgsql/repository/support/DefaultToolRepositoryInner.java
@@ -15,6 +15,7 @@ import modelengine.fitframework.annotation.Fit;
 import modelengine.fitframework.log.Logger;
 import modelengine.fitframework.serialization.ObjectSerializer;
 import modelengine.fitframework.transaction.Transactional;
+import modelengine.fitframework.util.CollectionUtils;
 import modelengine.jade.carver.tool.repository.pgsql.mapper.ToolMapper;
 import modelengine.jade.carver.tool.repository.pgsql.model.entity.ToolDo;
 import modelengine.jade.carver.tool.repository.pgsql.repository.ToolRepositoryInner;
@@ -81,6 +82,9 @@ public class DefaultToolRepositoryInner implements ToolRepositoryInner {
     @Override
     @Transactional
     public void deleteTools(List<String> uniqueNames) {
+        if (CollectionUtils.isEmpty(uniqueNames)) {
+            return;
+        }
         this.toolMapper.deleteTools(uniqueNames);
     }
